# docs/generated/openapi.rapidapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Naruto Mythos TCG Developer API",
    "description": "A developer-friendly REST API for querying the Naruto Mythos TCG catalogue. Card is a canonical identity; Printing is a collectible identity; Edition is scoped to a Set.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sets": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "List sets",
        "description": "Paginated list of sets. Supports filtering by language, legacy Set-level edition label and code, and sorting by name, release_date, or code. Collector Editions are listed under /v1/sets/{set_id}/editions.",
        "operationId": "list_sets_v1_sets_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page number, starting at 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Items per page, up to 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page, up to 100."
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Filter by language code (e.g. EN). Case-insensitive.",
              "title": "Language"
            },
            "description": "Filter by language code (e.g. EN). Case-insensitive."
          },
          {
            "name": "edition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Filter by edition. Case-insensitive.",
              "title": "Edition"
            },
            "description": "Filter by edition. Case-insensitive."
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Filter by set code. Case-insensitive.",
              "title": "Code"
            },
            "description": "Filter by set code. Case-insensitive."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field: name, release_date, or code.",
              "default": "release_date",
              "title": "Sort"
            },
            "description": "Sort field: name, release_date, or code."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order: asc or desc.",
              "default": "desc",
              "title": "Order"
            },
            "description": "Sort order: asc or desc."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSetsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sets/{public_id}": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "Get a single set",
        "description": "Look up a set by its public ID.",
        "operationId": "get_set_v1_sets__public_id__get",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetDetail"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Set not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SET_NOT_FOUND",
                    "message": "Set not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sets/{public_id}/cards": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "List a set's cards",
        "description": "Paginated list of cards belonging to one set. Supports sorting by number, name, or rarity. card_number is sorted lexically (as a string), never as a number.",
        "operationId": "list_set_cards_v1_sets__public_id__cards_get",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Page number, starting at 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number, starting at 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Items per page, up to 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Items per page, up to 100."
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field: number, name, or rarity.",
              "default": "number",
              "title": "Sort"
            },
            "description": "Sort field: number, name, or rarity."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order: asc or desc.",
              "default": "asc",
              "title": "Order"
            },
            "description": "Sort order: asc or desc."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCardsResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or sort parameters.",
            "content": {
              "application/json": {
                "examples": {
                  "pagination": {
                    "value": {
                      "error": {
                        "code": "INVALID_PAGINATION",
                        "message": "page must be >= 1."
                      }
                    }
                  },
                  "sort": {
                    "value": {
                      "error": {
                        "code": "INVALID_SORT",
                        "message": "Invalid sort field."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Set not found.",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SET_NOT_FOUND",
                    "message": "Set not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardImageResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "example": "front"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "example": "https://example.invalid/test-card.jpg"
          },
          "width": {
            "type": "integer",
            "nullable": true,
            "title": "Width",
            "example": 734
          },
          "height": {
            "type": "integer",
            "nullable": true,
            "title": "Height",
            "example": 1024
          }
        },
        "type": "object",
        "required": [
          "type",
          "url"
        ],
        "title": "CardImageResponse",
        "example": {
          "height": 1024,
          "type": "front",
          "url": "https://example.invalid/test-card.jpg",
          "width": 734
        }
      },
...
      "CardSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "TEST-001"
          },
          "number": {
            "type": "string",
            "title": "Number",
            "example": "001"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Character Alpha"
          },
          "subtitle": {
            "type": "string",
            "nullable": true,
            "title": "Subtitle",
            "example": null
          },
          "type": {
            "type": "string",
            "nullable": true,
            "title": "Type",
            "example": "Character"
          },
          "rarity": {
            "type": "string",
            "nullable": true,
            "title": "Rarity",
            "example": "Rare"
          },
          "set": {
            "$ref": "#/components/schemas/SetSummary"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CardImageResponse"
            },
            "type": "array",
            "title": "Images"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "name",
          "set"
        ],
        "title": "CardSummary",
        "example": {
          "id": "TEST-001",
          "images": [],
          "name": "Test Character Alpha",
          "number": "001",
          "rarity": "Rare",
          "set": {
            "code": "TST1",
            "edition": "1st Edition",
            "id": "test-set-1",
            "name": "Test Set Alpha"
          },
          "type": "Character"
        }
      },
...
      "EditionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "canonical_card_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Canonical Card Count"
          },
          "collectible_printing_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Collectible Printing Count"
          },
          "collector_reference_status": {
            "type": "string",
            "pattern": "^(VERIFIED|NONE)$",
            "title": "Collector Reference Status"
          },
          "collector_reference_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Collector Reference Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "canonical_card_count",
          "collectible_printing_count",
          "collector_reference_status",
          "collector_reference_count"
        ],
        "title": "EditionSummary"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "CARD_NOT_FOUND",
          "SET_NOT_FOUND",
          "EDITION_NOT_FOUND",
          "PRINTING_NOT_FOUND",
          "COLLECTOR_REFERENCE_NOT_FOUND",
          "KEYWORD_NOT_FOUND",
          "INVALID_FILTER",
          "INVALID_PAGINATION",
          "INVALID_SORT",
          "FORBIDDEN",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "Known/expected error codes for the /v1 API.\n\nThis list is expected to grow in later phases as route handlers are\nimplemented (e.g. Phase 5+ will need to actually raise CARD_NOT_FOUND\nwhen a lookup misses) -- growing it is an additive change."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedCardsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CardSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedCardsResponse",
        "example": {
          "data": [],
          "pagination": {
            "has_next": false,
            "has_previous": false,
            "limit": 50,
            "page": 1,
            "pages": 0,
            "total": 0
          }
        }
      },
...
      "PaginatedSetsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SetDetail"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedSetsResponse",
        "description": "Paginated GET /v1/sets response.\n\nUses `SetDetail` (the full representation), not a further-trimmed-down\nsummary, as the list item shape. Decision, documented per the brief:\nunlike cards -- which can number in the thousands and are paginated\nspecifically to keep list responses lightweight -- the number of sets\nin a TCG catalogue is small (tens, not thousands), so there is no\nmeaningful payload-size concern with returning full detail per set even\nin a list response. Giving developers the complete set record (release\ndate, printed totals, etc.) directly from the list endpoint avoids an\nextra round trip to `GET /v1/sets/{id}` for data they'll almost always\nwant anyway when browsing sets.",
        "example": {
          "data": [],
          "pagination": {
            "has_next": false,
            "has_previous": false,
            "limit": 50,
            "page": 1,
            "pages": 0,
            "total": 0
          }
        }
      },
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "example": 50
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "example": 630
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "example": 13
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "example": true
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous",
            "example": false
          }
        },
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta",
        "example": {
          "has_next": true,
          "has_previous": false,
          "limit": 50,
          "page": 1,
          "pages": 13,
          "total": 630
        }
      },
...
      "SetDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "test-set-1"
          },
          "code": {
            "type": "string",
            "nullable": true,
            "title": "Code",
            "example": "TST1"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Set Alpha"
          },
          "edition": {
            "type": "string",
            "nullable": true,
            "title": "Edition",
            "example": "1st Edition"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "example": "EN"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "title": "Release Date",
            "example": "2026-01-01"
          },
          "printed_total": {
            "type": "integer",
            "nullable": true,
            "title": "Printed Total",
            "example": 100
          },
          "total_with_variants": {
            "type": "integer",
            "nullable": true,
            "title": "Total With Variants",
            "example": 150
          },
          "logo_url": {
            "type": "string",
            "nullable": true,
            "title": "Logo Url",
            "example": null
          },
          "symbol_url": {
            "type": "string",
            "nullable": true,
            "title": "Symbol Url",
            "example": null
          },
          "canonical_card_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Canonical Card Count",
            "default": 0
          },
          "collectible_printing_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Collectible Printing Count",
            "default": 0
          },
          "editions": {
            "items": {
              "$ref": "#/components/schemas/EditionSummary"
            },
            "type": "array",
            "title": "Editions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "language"
        ],
        "title": "SetDetail",
        "example": {
          "canonical_card_count": 0,
          "code": "TST1",
          "collectible_printing_count": 0,
          "edition": "1st Edition",
          "editions": [],
          "id": "test-set-1",
          "language": "EN",
          "name": "Test Set Alpha",
          "printed_total": 100,
          "release_date": "2026-01-01",
          "total_with_variants": 150
        }
      },
...
      "SetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "test-set-1"
          },
          "code": {
            "type": "string",
            "nullable": true,
            "title": "Code",
            "example": "TST1"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Test Set Alpha"
          },
          "edition": {
            "type": "string",
            "nullable": true,
            "title": "Edition",
            "example": "1st Edition"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "SetSummary",
        "example": {
          "code": "TST1",
          "edition": "1st Edition",
          "id": "test-set-1",
          "name": "Test Set Alpha"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}